refactor(svelte2tsx): extract ComponentEvents into script/component_events.rs
Pure move: no logic changes, only `use` additions and `pub(super)` visibility
adjustments. The corresponding unit tests move with the code.
